+++ CHANGELOG.md
@@ -2,9 +2,9 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-* Replace color gradiants
-* Nag component stateChanged -> stateChange
-* Fixed disable component colors
+* Feature: Add new color gradiants
+* Fix: Nag component stateChanged -> stateChange
+* Feature: Fixed disable component colors
 * Feature: Added light and light italic fonts
 
 --------------------